backend/requirements.txt
flask
numpy
pandas
scikit-learn
joblib
xgboost
tensorflow
matplotlib
seaborn
pycryptodomex
flask_cors